requirements.txt
colorama~=0.4.6
python-dotenv~=1.0.0
DateTime~=5.1
Pillow~=9.5.0
noise~=1.2.2
numpy~=1.24.3
scipy~=1.10.1
pygame~=2.4.0